+++ test/integration/kubernetes_deploy_test.rb
@@ -1068,6 +1068,6 @@ unknown field \"myKey\" in io.k8s.apimachinery.pkg.apis.meta.v1.ObjectMeta",
     skip if KUBE_SERVER_VERSION < Gem::Version.new('1.8.0')
     assert_deploy_success(deploy_fixtures("hpa"))
     assert_deploy_success(deploy_fixtures("hpa", subset: ["deployment.yml"]))
-    assert_logs_match_all(['The following resources were pruned: horizontalpodautoscaler "hello-hpa"'])
+    assert_logs_match_all([/The following resources were pruned: horizontalpodautoscaler(.autoscaling)? "hello-hpa"/])
   end
 end